Update blobify() of all distillers, take Distiller.self instead of `url`
Distiller plugins will benefit for more context and a proper
"environment" around them, rather than acting purely in isolation with
only the URL that they've been passed.  These changes pass in the
context of the whole Distiller instance.

Specifically, this will allow distillers to receive auth credentials and
access to the indexing queue instead of finding it for themselves.