Fix version tag display in frontend when umad.py is run from WSGI
WSGI is, of course, not the same as running as __main__ from the command
line. It's not efficient, but I've had the "main" search() function look
for a version tag file and read it at the start of every request.

Ideally this would be done earlier, in the app container, but I'm not
sure if that's possible. Either way, opening a file and reading one line
on every request shouldn't exactly be an impossible load.