Don't verify SSL cert in testing script
This is an annoyance, but Python2's support for SNI is lacklustre. In an
SNI deployment, the Requests library will get confused if it gets a
hostname-to-certificate mismatch, which this change dodges. Less of a
problem if you have a handy wildcard cert, but you may not have one for
your staging/testing subdomain.