Add ability to disable/vary SSL verification when hitting indexer
Our staging environment unfortunately doesn't have a legit SSL
certificate, so we need to either pass it a CA to verify against or
disable verification. This change allows that.

The default in the Requests library is True, and you can set the
UMAD_INDEXER_VERIFY_SSL envvar to override that. The strings "True" and
"False" are coerced to the proper python boolean type, while other
strings will be passed through. I've no idea what an empty string will
do, hmm...