Dynamically build the list of Distiller classes
This saves on repetition, which would be prone to error. We already
explicitly import each Distiller in distil/__init__.py, so we have some
control over what appears in scope.

This method is probably technically undefined, but appears to leave the
`distillers` list in alphabetic order. This should not, but cannot be
guaranteed not to cause problems when we try to find a distiller for a
given URL. We don't expect any problems though, as we probably shouldn't
have more than one distiller that can handle a particular URL.