Remove an insufficient check from the frontend heartbeat
Checking that there's at least MAX_HITS results for each index is not a
good idea. It's a fair assertion, but there may be time when an index
explodes and disappears. That's not a *frontend* problem, so it's not
grounds for the load balancer to have a fit. We'll address that with a
separate Nagios check.