RUNNING.md isn't really accurate nor relevant any more, nuke it.
I used to be a simple app like you, then I took an arrow in the knee and
split into seven loosely coupled mini services. Which is great for
scalability and resilience, but not so good for testing.

You can certainly run the whole lot on your workstation without undue
difficulty, with a bunch of terminals and passing of environment
variables. I might get to write some solid testing docs in future, but
let's get rid of the immediate cruft.