Deal with Gollum being weird about where it POSTs stuff to
The idea was nice, catch POST events in the webserver's access logs. The
truth of the matter, though, is that Gollum doesn't POST to meaningful
endpoints. As such, those URLs can't be used.

Examples:

https://docs.anchor.net.au/repo/root/create
https://docs.anchor.net.au/repo/root/edit/What-am-I-doing-here

We could attempt to detect Create and Edit actions in the URL, but it's
too brittle and prone to failure, particularly because we don't know the
repo root at this point.

Instead, hardcode our handling for Anchor docs and eschew the use of
POST detection. This isn't so bad, seeing as Gollum needs special
handling anyway (commits can be pushed straight to the repo in addition
to the edits through the web interface).

Now we look for INDEX events in the logs. INDEX will either be
synthesised in the repo's post-receive hook, or in Gollum's config.ru
with a post-commit hook.